Change the counters to be all the same type -- u_long.

+++ sys/netinet/udp_var.h
@@ -31,7 +31,7 @@
  * SUCH DAMAGE.
  *
  *	from: @(#)udp_var.h	7.7 (Berkeley) 6/28/90
- *	$Id: udp_var.h,v 1.4 1994/01/08 23:17:20 mycroft Exp $
+ *	$Id: udp_var.h,v 1.5 1994/01/10 23:27:48 mycroft Exp $
  */
 
 #ifndef _NETINET_UDP_VAR_H_
@@ -58,16 +58,16 @@ struct	udpiphdr {
 
 struct	udpstat {
 				/* input statistics: */
-	int	udps_ipackets;		/* total input packets */
-	int	udps_hdrops;		/* packet shorter than header */
-	int	udps_badsum;		/* checksum error */
-	int	udps_badlen;		/* data length larger than packet */
-	int	udps_noport;		/* no socket on port */
-	int	udps_noportbcast;	/* of above, arrived as broadcast */
-	int	udps_fullsock;		/* not delivered, input socket full */
-	int	udpps_pcbcachemiss;	/* input packets missing pcb cache */
+	u_long	udps_ipackets;		/* total input packets */
+	u_long	udps_hdrops;		/* packet shorter than header */
+	u_long	udps_badsum;		/* checksum error */
+	u_long	udps_badlen;		/* data length larger than packet */
+	u_long	udps_noport;		/* no socket on port */
+	u_long	udps_noportbcast;	/* of above, arrived as broadcast */
+	u_long	udps_fullsock;		/* not delivered, input socket full */
+	u_long	udpps_pcbcachemiss;	/* input packets missing pcb cache */
 				/* output statistics: */
-	int	udps_opackets;		/* total output packets */
+	u_long	udps_opackets;		/* total output packets */
 };
 
 #define	UDP_TTL		30	/* default time to live for UDP packets */
